backend/_config/requirements/development.txt
# backend/_config/requirements/development.txt

-r base.txt

# Development tools
ipython==8.25.0
django-extensions==3.2.3
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
black==24.8.0
flake8==7.0.0
isort==5.13.2